backend/requirements.txt
Django==3.2.7
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
django-cors-headers==3.13.0
gunicorn==20.0.4
psycopg2-binary==2.9.3
